+++ CodeTester-IntelliJ-IDEA/src/main/java/de/fxnm/toolwindow/CodeTesterToolWindowFactory.java
@@ -1,6 +1,8 @@
 package de.fxnm.toolwindow;
 
+import com.intellij.openapi.components.ServiceManager;
 import com.intellij.openapi.project.Project;
+import com.intellij.openapi.util.Pair;
 import com.intellij.openapi.wm.ToolWindow;
 import com.intellij.openapi.wm.ToolWindowFactory;
 import com.intellij.openapi.wm.ToolWindowType;
@@ -17,20 +19,9 @@ public class CodeTesterToolWindowFactory implements ToolWindowFactory {
 
     private static final int MAX_LENGTH = 10;
 
-    public static void createResultToolWindow(final @NotNull ToolWindow toolWindow,
-                                              final ResultTreeNode resultTreeNode) {
-        final Content content = toolWindow.getContentManager().getFactory().createContent(
-                new ResultToolWindowPanel(resultTreeNode).getComponent(),
-                restrictStringLength(MAX_LENGTH, resultTreeNode.getCheck().getCheckName()),
-                true);
-        content.putUserData(ToolWindow.SHOW_CONTENT_ICON, Boolean.TRUE);
-
-        content.setIcon(resultTreeNode.getIcon());
-
-        toolWindow.getContentManager().addContent(content);
-        toolWindow.getContentManager().setSelectedContent(content);
+    public static CodeTesterToolWindowFactory getService(final Project project) {
+        return ServiceManager.getService(project, CodeTesterToolWindowFactory.class);
     }
-
 
     private static String restrictStringLength(final int maxLength, final String string) {
         if (string.length() <= maxLength) {
@@ -47,6 +38,26 @@ public class CodeTesterToolWindowFactory implements ToolWindowFactory {
         stringBuilder.append("...");
 
         return stringBuilder.toString();
+    }
+
+    public Pair<Content, ResultToolWindowPanel> createResultToolWindow(final @NotNull ToolWindow toolWindow,
+                                                                       final @NotNull ResultTreeNode resultTreeNode) {
+
+        final ResultToolWindowPanel resultToolWindowPanel = new ResultToolWindowPanel(resultTreeNode);
+
+        final Content content = toolWindow.getContentManager().getFactory().createContent(
+                resultToolWindowPanel.getComponent(),
+                restrictStringLength(MAX_LENGTH, resultTreeNode.getCheck().getCheckName()),
+                true);
+
+        content.putUserData(ToolWindow.SHOW_CONTENT_ICON, Boolean.TRUE);
+        content.setIcon(resultTreeNode.getIcon());
+        content.setDescription(resultTreeNode.getCheck().getCheckName());
+
+        toolWindow.getContentManager().addContent(content);
+        toolWindow.getContentManager().setSelectedContent(content);
+
+        return new Pair<>(content, resultToolWindowPanel);
     }
 
     @Override